% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDrivingRoute.R
\name{addDrivingRoute}
\alias{addDrivingRoute}
\title{Add a layer of driving route to a Leaflet map widget}
\usage{
addDrivingRoute(map, origin = c(116.379028, 39.865042),
  destination = c(116.427281, 39.903719), waypoints = NULL, strategy = 0,
  key, layerId = NULL, group = NULL, weight = 5, opacity = 1,
  popup = NULL)
}
\arguments{
\item{map}{a leaflet map widget}

\item{origin}{a numeric vector defining the origin location, format c(longitude, latitude)}

\item{destination}{a numeric vector defining the destination, format c(longitude, latitude)}

\item{waypoints}{a data frame consists of two columns, where the first column is longitude
and the second latitude, at most 16 rows.}

\item{strategy}{refer to the original api http://lbs.amap.com/api/webservice/reference/direction/#t7}

\item{key}{access key of amap api, please register at the website if you do not have one.}

\item{layerId}{layerId in leaflet}

\item{group}{group in leaflet}

\item{weight}{weight in leaflet}

\item{opacity}{opacity in leaflet}

\item{popup}{popup in leaflet, available attributes include status}
}
\value{
a leaflet map widget
}
\description{
This function adds a layer of driving route to a Leaflet map widget using \pkg{leaflet}.
}
\examples{
\dontrun{
leaflet() \%>\% addDrivingRoute(key="your key from amap website")
}
}

