% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRoadMap.R
\name{addRoadMap}
\alias{addRoadMap}
\title{Add a layer of road map to a Leaflet map widget}
\usage{
addRoadMap(map, center = c(116.4, 39.9), zoom = 10, layerId = NULL,
  group = NULL)
}
\arguments{
\item{map}{a leaflet map widget}

\item{center}{a numeric vector defineing the center point of the map widget,
default set to Beijing Tiananmen. Notice that the longitude and latitude
should follow the Chinese standard coordinate GCJ-02}

\item{zoom}{zoom level of the 10, default set to 10}

\item{layerId}{layerId in leaflet}

\item{group}{group in leaflet}
}
\value{
a leaflet map widget
}
\description{
Add a layer of road map to a Leaflet map widget
}
\examples{
leaflet() \%>\% addRoadMap()
}

